+++ composenav/src/main/java/com/adamkobus/compose/navigation/data/GlobalGraph.kt
@@ -3,7 +3,7 @@ package com.adamkobus.compose.navigation.data
 import com.adamkobus.compose.navigation.destination.NavDestination
 import com.adamkobus.compose.navigation.destination.popDestination
 
-object GlobalGraph : NavGraph("__global__") {
+object GlobalGraph : NavGraph(name = "__global__", reservedNameCheck = false) {
 
     private val Root = navDestination("__root__")