+++ composenav/src/main/java/com/adamkobus/compose/navigation/error/ReservedNameError.kt
@@ -1,0 +1,3 @@
+package com.adamkobus.compose.navigation.error
+
+class ReservedNameError(message: String) : RuntimeException(message)